biodata/airflow/deployment/trigger-service/requirements.txt
functions-framework==3.*
google-auth==2.*
google-cloud-orchestration-airflow==1.*
requests==2.*